fix font awesome icons
