+++ Dashboard/src/Components/TopBar/UserProfileButton.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement, FunctionComponent } from 'react';
 import MenuItem from 'CommonUI/src/Components/Basic/Button/MenuButton/MenuItem';
 import UserInfo from 'CommonUI/src/Components/Dashboard/TopBar/UserProfile/UserInfo';
 import { IconProp } from 'CommonUI/src/Components/Basic/Icon/Icon';
-import IconButton from 'CommonUI/src/Components/Basic/Button/IconButton';
+import IconButton from 'CommonUI/src/Components/Basic/Button/MenuButton/IconButton';
 
 const CreateButton: FunctionComponent = (): ReactElement => {
     return (
